+++ node/libs/roles/src/validator/messages/block.rs
@@ -1,8 +1,45 @@
 //! Messages related to blocks.
-
-use super::{CommitQC, CommitQCVerifyError};
+use super::v1::FinalBlock;
 use std::fmt;
 use zksync_consensus_crypto::{keccak256::Keccak256, ByteFmt, Text, TextFmt};
+/// Represents a blockchain block across different consensus protocol versions (including pre-genesis blocks).
+#[derive(Debug, Clone, PartialEq)]
+pub enum Block {
+    /// Block with number `<genesis.first_block`.
+    PreGenesis(PreGenesisBlock),
+    /// Block with number `>=genesis.first_block`.
+    Final(FinalBlock),
+}
+
+impl From<PreGenesisBlock> for Block {
+    fn from(b: PreGenesisBlock) -> Self {
+        Self::PreGenesis(b)
+    }
+}
+
+impl From<FinalBlock> for Block {
+    fn from(b: FinalBlock) -> Self {
+        Self::Final(b)
+    }
+}
+
+impl Block {
+    /// Block number.
+    pub fn number(&self) -> BlockNumber {
+        match self {
+            Self::PreGenesis(b) => b.number,
+            Self::Final(b) => b.number(),
+        }
+    }
+
+    /// Payload.
+    pub fn payload(&self) -> &Payload {
+        match self {
+            Self::PreGenesis(b) => &b.payload,
+            Self::Final(b) => &b.payload,
+        }
+    }
+}
 
 /// Payload of the block. Consensus algorithm does not interpret the payload
 /// (except for imposing a size limit for the payload). Proposing a payload
@@ -92,95 +129,6 @@ impl fmt::Display for BlockNumber {
     }
 }
 
-/// A block header.
-/// WARNING: any change to this struct may invalidate preexisting signatures. See `TimeoutQC` docs.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct BlockHeader {
-    /// Number of the block.
-    pub number: BlockNumber,
-    /// Payload of the block.
-    pub payload: PayloadHash,
-}
-
-/// A block that has been finalized by the consensus protocol.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct FinalBlock {
-    /// Payload of the block. Should match `header.payload` hash.
-    pub payload: Payload,
-    /// Justification for the block. What guarantees that the block is final.
-    pub justification: CommitQC,
-}
-
-impl FinalBlock {
-    /// Creates a new finalized block.
-    pub fn new(payload: Payload, justification: CommitQC) -> Self {
-        assert_eq!(justification.header().payload, payload.hash());
-
-        Self {
-            payload,
-            justification,
-        }
-    }
-
-    /// Header of the block.
-    pub fn header(&self) -> &BlockHeader {
-        &self.justification.message.proposal
-    }
-
-    /// Number of the block.
-    pub fn number(&self) -> BlockNumber {
-        self.header().number
-    }
-
-    /// Verifies internal consistency of this block.
-    pub fn verify(&self, genesis: &super::Genesis) -> Result<(), BlockValidationError> {
-        let payload_hash = self.payload.hash();
-        if payload_hash != self.header().payload {
-            return Err(BlockValidationError::HashMismatch {
-                header_hash: self.header().payload,
-                payload_hash,
-            });
-        }
-        self.justification
-            .verify(genesis)
-            .map_err(BlockValidationError::Justification)
-    }
-}
-
-impl ByteFmt for FinalBlock {
-    fn decode(bytes: &[u8]) -> anyhow::Result<Self> {
-        zksync_protobuf::decode(bytes)
-    }
-
-    fn encode(&self) -> Vec<u8> {
-        zksync_protobuf::encode(self)
-    }
-}
-
-/// Errors that can occur validating a `FinalBlock` received from a node.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum BlockValidationError {
-    /// Block payload doesn't match the block header.
-    #[error(
-        "block payload doesn't match the block header (hash in header: {header_hash:?}, \
-             payload hash: {payload_hash:?})"
-    )]
-    HashMismatch {
-        /// Payload hash in block header.
-        header_hash: PayloadHash,
-        /// Hash of the payload.
-        payload_hash: PayloadHash,
-    },
-    /// Failed verifying quorum certificate.
-    #[error("failed verifying quorum certificate: {0:#?}")]
-    Justification(#[source] CommitQCVerifyError),
-}
-
-/// TODO: docs
-#[derive(Debug, Clone, PartialEq)]
-pub struct Justification(pub Vec<u8>);
-
 /// Block before `genesis.first_block`
 /// with an external (non-consensus) justification.
 #[derive(Debug, Clone, PartialEq)]
@@ -195,39 +143,4 @@ pub struct PreGenesisBlock {
 
 /// TODO: docs
 #[derive(Debug, Clone, PartialEq)]
-pub enum Block {
-    /// Block with number `<genesis.first_block`.
-    PreGenesis(PreGenesisBlock),
-    /// Block with number `>=genesis.first_block`.
-    Final(FinalBlock),
-}
-
-impl From<PreGenesisBlock> for Block {
-    fn from(b: PreGenesisBlock) -> Self {
-        Self::PreGenesis(b)
-    }
-}
-
-impl From<FinalBlock> for Block {
-    fn from(b: FinalBlock) -> Self {
-        Self::Final(b)
-    }
-}
-
-impl Block {
-    /// Block number.
-    pub fn number(&self) -> BlockNumber {
-        match self {
-            Self::PreGenesis(b) => b.number,
-            Self::Final(b) => b.number(),
-        }
-    }
-
-    /// Payload.
-    pub fn payload(&self) -> &Payload {
-        match self {
-            Self::PreGenesis(b) => &b.payload,
-            Self::Final(b) => &b.payload,
-        }
-    }
-}
+pub struct Justification(pub Vec<u8>);
