+++ node/libs/roles/src/validator/messages/mod.rs
@@ -1,25 +1,15 @@
 //! Messages exchanged between validators.
 
 mod block;
-mod committee;
-mod consensus;
 mod discovery;
 mod genesis;
-mod leader_proposal;
 mod msg;
-mod replica_commit;
-mod replica_new_view;
-mod replica_timeout;
+mod testonly;
 #[cfg(test)]
 mod tests;
+pub mod v1;
 
 pub use block::*;
-pub use committee::*;
-pub use consensus::*;
 pub use discovery::*;
 pub use genesis::*;
-pub use leader_proposal::*;
 pub use msg::*;
-pub use replica_commit::*;
-pub use replica_new_view::*;
-pub use replica_timeout::*;
