+++ node/libs/roles/src/validator/messages/v1/leader_proposal.rs
@@ -1,7 +1,6 @@
-use super::{
-    BlockNumber, CommitQC, CommitQCVerifyError, Genesis, Payload, PayloadHash, TimeoutQC,
-    TimeoutQCVerifyError, View,
-};
+use crate::validator::{BlockNumber, Genesis, Payload, PayloadHash};
+
+use super::{CommitQC, CommitQCVerifyError, TimeoutQC, TimeoutQCVerifyError, View};
 
 /// A proposal message from the leader.
 #[derive(Clone, Debug, PartialEq, Eq)]
