+++ node/libs/roles/src/validator/messages/v1/tests/versions.rs
@@ -1,4 +1,3 @@
-use super::*;
 use anyhow::Context as _;
 use zksync_consensus_crypto::Text;
